refactor(canvas): extract skia_canvas_shaders.cpp (R2-3 fifth slice)
Phase 4 R2-3 follow-up. Pull 474 lines covering 4 adjacent
GPU/SkSL-shader concerns out of skia_canvas.cpp into one sibling TU:

- GPU SDF Shape Primitives — kSDFShapeSkSL runtime effect + the
  draw_sdf_shape / draw_sdf_ring / draw_sdf_rect family
- Custom SkSL shader rendering — draw_with_shader() compiles a
  user-supplied SkSL program with caller-bound uniforms
- Blur backdrop — draw_blurred_backdrop() composites an
  SkImageFilters::Blur saveLayer with an optional tint
- GPU Waveform — kWaveformSkSL runtime effect + draw_waveform() /
  draw_spectrum_bars() for shader-driven 1D-texture rendering

- skia_canvas.cpp: 3,348 → 2,873 lines (-475 net).
- skia_canvas_shaders.cpp: 530 lines.

Tests:
- pulp-canvas builds clean on macOS arm64 (Debug, GPU + Skia).
- pulp-test-canvas: 1,550 assertions / 63 cases — pass.

R2-3 progress (within this PR + mask in prior PRs):
  mask + gradients + box-shadow + opacity + shaders (5 slices done).
skia_canvas.cpp went 3,943 (this PR start) → 2,873 = -1,070 lines, -27%.
